webrtc/RTCPeerConnection-setRemoteDescription-rollback.html

<!doctype html>
<meta charset=utf-8>
<title>RTCPeerConnection.prototype.setRemoteDescription rollback</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="RTCPeerConnection-helper.js"></script>
<script>
  'use strict';

  // Test is based on the following editor draft:
  // https://w3c.github.io/webrtc-pc/archives/20170605/webrtc.html

  // The following helper functions are called from RTCPeerConnection-helper.js:
  //   assert_session_desc_similar
  //   generateAudioReceiveOnlyOffer
  //   generateDataChannelOffer

  /*
    4.3.2.  Interface Definition
      [Constructor(optional RTCConfiguration configuration)]
      interface RTCPeerConnection : EventTarget {
        Promise<void>                      setLocalDescription(
            RTCSessionDescriptionInit description);

        readonly attribute RTCSessionDescription? localDescription;
        readonly attribute RTCSessionDescription? currentLocalDescription;
        readonly attribute RTCSessionDescription? pendingLocalDescription;

        Promise<void>                      setRemoteDescription(
            RTCSessionDescriptionInit description);

        readonly attribute RTCSessionDescription? remoteDescription;
        readonly attribute RTCSessionDescription? currentRemoteDescription;
        readonly attribute RTCSessionDescription? pendingRemoteDescription;
        ...
      };

    4.6.2.  RTCSessionDescription Class
      dictionary RTCSessionDescriptionInit {
        required RTCSdpType type;
                 DOMString  sdp = "";
      };

    4.6.1.  RTCSdpType
      enum RTCSdpType {
        "offer",
        "pranswer",
        "answer",
        "rollback"
      };
   */

  /*
    4.3.1.6.  Set the RTCSessionSessionDescription
      2.2.3.  Otherwise, if description is set as a remote description, then run one
              of the following steps:
        - If description is of type "rollback", then this is a rollback.
          Set connection.pendingRemoteDescription to null and signaling state to stable.
   */
  promise_test(t => {
    const pc = new RTCPeerConnection();
    t.add_cleanup(() => pc.close());

    const states = [];
    pc.addEventListener('signalingstatechange', () => states.push(pc.signalingState));

    return generateDataChannelOffer(pc)
    .then(offer => pc.setRemoteDescription(offer))
    .then(() => {
      assert_equals(pc.signalingState, 'have-remote-offer');
      assert_not_equals(pc.remoteDescription, null);
      assert_not_equals(pc.pendingRemoteDescription, null);
      assert_equals(pc.currentRemoteDescription, null);

      return pc.setRemoteDescription({type: 'rollback'});
    })
    .then(() => {
      assert_equals(pc.signalingState, 'stable');
      assert_equals(pc.remoteDescription, null);
      assert_equals(pc.pendingRemoteDescription, null);
      assert_equals(pc.currentRemoteDescription, null);

      assert_array_equals(states, ['have-remote-offer', 'stable']);
    });
  }, 'setRemoteDescription(rollback) in have-remote-offer state should revert to stable state');

  /*
    4.3.1.6.  Set the RTCSessionSessionDescription
      2.3.  If the description's type is invalid for the current signaling state of
            connection, then reject p with a newly created InvalidStateError and abort
            these steps.

    [jsep]
      4.1.8.2.  Rollback
        - Rollback can only be used to cancel proposed changes;
          there is no support for rolling back from a stable state to a
          previous stable state
   */
  promise_test(t => {
    const pc = new RTCPeerConnection();
    t.add_cleanup(() => pc.close());
    return promise_rejects(t, 'InvalidStateError',
      pc.setRemoteDescription({type: 'rollback'}));
  }, `setRemoteDescription(rollback) from stable state should reject with InvalidStateError`);

  promise_test(t => {
    const pc = new RTCPeerConnection();
    t.add_cleanup(() => pc.close());
    return generateAudioReceiveOnlyOffer(pc)
    .then(offer => pc.setRemoteDescription(offer))
    .then(() => pc.setRemoteDescription({
      type: 'rollback',
      sdp: '!<Invalid SDP Content>;'
    }));
  }, `setRemoteDescription(rollback) should ignore invalid sdp content and succeed`);

  promise_test(async t => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection();
    t.add_cleanup(() => pc1.close());
    t.add_cleanup(() => pc2.close());

    // We don't use this right away
    const offer1 = await pc1.createOffer({offerToReceiveAudio: true});

    // Create offer from pc2, apply and rollback on pc1
    const offer2 = await pc2.createOffer({offerToReceiveAudio: true});
    await pc1.setRemoteDescription(offer2);
    await pc1.setRemoteDescription({type: "rollback"});

    // Then try applying pc1's old offer
    await pc1.setLocalDescription(offer1);
  }, `local offer created before setRemoteDescription(remote offer) then rollback should still be usable`);

  promise_test(async t => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection();
    t.add_cleanup(() => pc1.close());
    t.add_cleanup(() => pc2.close());

    const stream = await getNoiseStream({video: true});
    t.add_cleanup(() => stream.getTracks().forEach(track => track.stop()));
    pc1.addTrack(stream.getTracks()[0], stream);

    // We don't use this right away. pc1 has provisionally decided that the
    // (only) transceiver is bound to level 0.
    const offer1 = await pc1.createOffer();

    // Create offer from pc2, apply and rollback on pc1
    const offer2 = await pc2.createOffer({offerToReceiveAudio: true, offerToReceiveVideo: true});
    // pc1 now should change its mind about what level its video transceiver is
    // bound to. It was 0, now it is 1.
    await pc1.setRemoteDescription(offer2);

    // Rolling back should put things back the way they were.
    await pc1.setRemoteDescription({type: "rollback"});

    // Then try applying pc1's old offer
    await pc1.setLocalDescription(offer1);
  }, "local offer created before setRemoteDescription(remote offer) with different transceiver level assignments then rollback should still be usable");

  promise_test(async t => {
    const pc1 = new RTCPeerConnection();
    t.add_cleanup(() => pc1.close());
    const pc2 = new RTCPeerConnection();
    t.add_cleanup(() => pc2.close());

    const stream = await getNoiseStream({video: true});
    t.add_cleanup(() => stream.getTracks().forEach(track => track.stop()));
    pc1.addTrack(stream.getTracks()[0], stream);

    await pc2.setRemoteDescription(await pc1.createOffer());
    assert_equals(pc2.getTransceivers().length, 1);
    await pc2.setRemoteDescription({type: "rollback"});
    // assert_equals(pc2.getTransceivers().length, 0); // Decreasing number of receivers is not observable from JS.
  }, "Rollback remote offer with removing transceiver");

  promise_test(async t => {
    const pc = new RTCPeerConnection();
    t.add_cleanup(() => pc.close());

    const stream = await getNoiseStream({video: true});
    t.add_cleanup(() => stream.getTracks().forEach(track => track.stop()));
    pc.addTrack(stream.getTracks()[0], stream);

    const states = [];
    const signalingstatechangeResolver = new Resolver();
    pc.onsignalingstatechange = () => {
      states.push(pc.signalingState);
      signalingstatechangeResolver.resolve();
    };

    const offer = await pc.createOffer();
    await pc.setLocalDescription(offer);
    await pc.setRemoteDescription({type: "rollback"});
    await pc.setLocalDescription(offer);
    await signalingstatechangeResolver.promise;
    assert_array_equals(states, ['have-local-offer', 'stable', 'have-local-offer']);
  }, "Rollback local offer explicitly");

  promise_test(async t => {
    const pc1 = new RTCPeerConnection();
    t.add_cleanup(() => pc1.close());
    const pc2 = new RTCPeerConnection();
    t.add_cleanup(() => pc2.close());

    const states = [];
    const signalingstatechangeResolver = new Resolver();
    pc1.onsignalingstatechange = () => {
      states.push(pc1.signalingState);
      signalingstatechangeResolver.resolve();
    };
    const stream1 = await getNoiseStream({audio: true});
    t.add_cleanup(() => stream1.getTracks().forEach(track => track.stop()));
    pc1.addTrack(stream1.getTracks()[0], stream1);

    const stream2 = await getNoiseStream({video: true});
    t.add_cleanup(() => stream2.getTracks().forEach(track => track.stop()));
    pc2.addTrack(stream2.getTracks()[0], stream2);

    await pc1.setLocalDescription(await pc1.createOffer());
    var onnegotiationneededCount = 0;
    const negotiationneededResolver = new Resolver();
    pc1.onnegotiationneeded = () => {
      onnegotiationneededCount += 1;
      negotiationneededResolver.resolve();
    };
    await pc1.setRemoteDescription(await pc2.createOffer());
    await pc1.setLocalDescription(await pc1.createAnswer());
    await negotiationneededResolver.promise;
    assert_equals(onnegotiationneededCount, 1);
    await signalingstatechangeResolver.promise;
    // For some reason stable state as part of implicit rollback is not observable from JS.
    assert_array_equals(states, ['have-local-offer', 'have-remote-offer', 'stable']);
  }, "Rollback local offer implicitly visiting stable state in between. At the end fire negotiationneeded due to rollback");

  promise_test(async t => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection();
    t.add_cleanup(() => pc1.close());
    t.add_cleanup(() => pc2.close());

    const stream = await getNoiseStream({audio: true});
    t.add_cleanup(() => stream.getTracks().forEach(track => track.stop()));
    pc1.addTrack(stream.getTracks()[0], stream);

    await pc1.setLocalDescription(await pc1.createOffer());
    await pc2.setRemoteDescription(pc1.pendingLocalDescription);

    await pc2.setLocalDescription(await pc2.createAnswer());
    await pc1.setRemoteDescription(pc2.localDescription);

    // In stable state add video on both end and make sure video transceiver is not killed.

    const stream1 = await getNoiseStream({video: true});
    t.add_cleanup(() => stream1.getTracks().forEach(track => track.stop()));
    pc1.addTrack(stream1.getTracks()[0], stream1);
    await pc1.setLocalDescription(await pc1.createOffer());

    const stream2 = await getNoiseStream({video: true});
    t.add_cleanup(() => stream2.getTracks().forEach(track => track.stop()));
    pc2.addTrack(stream2.getTracks()[0], stream2);
    const offer2 = await pc2.createOffer();
    await pc2.setRemoteDescription(pc1.pendingLocalDescription);
    await pc2.setRemoteDescription({type: "rollback"});
    assert_equals(pc2.getTransceivers().length, 2);
    await pc2.setLocalDescription(offer2);
  },"Rollback remote offer to negotiated stable state and then apply local offer");

promise_test(async t => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection();
    t.add_cleanup(() => pc1.close());
    t.add_cleanup(() => pc2.close());

    const stream1 = await getNoiseStream({audio: true});
    t.add_cleanup(() => stream1.getTracks().forEach(track => track.stop()));
    var sender = pc1.addTrack(stream1.getTracks()[0], stream1);

    const stream2 = await getNoiseStream({audio: true});
    t.add_cleanup(() => stream2.getTracks().forEach(track => track.stop()));
    pc2.addTrack(stream2.getTracks()[0], stream2);

    await pc1.setLocalDescription(await pc1.createOffer());
    await pc2.setRemoteDescription(pc1.pendingLocalDescription);

    await pc2.setLocalDescription(await pc2.createAnswer());
    await pc1.setRemoteDescription(pc2.localDescription);

    // In stable state make audio recvonly and then rollback.

    pc1.removeTrack(sender);
    await pc1.setLocalDescription(await pc1.createOffer());
    assert_equals(pc2.getTransceivers()[0].direction, "sendrecv");
    await pc2.setRemoteDescription(pc1.pendingLocalDescription);
    assert_equals(pc2.getTransceivers()[0].direction, "sendrecv");
    await pc2.setRemoteDescription({type: "rollback"});
    assert_equals(pc2.getTransceivers().length, 1);
 //   assert_equals(pc2.getTransceivers()[0].direction, "sendonly");
  },"Rollback remote offer to negotiated stable state and then apply local offer");

  promise_test(async t => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection();
    t.add_cleanup(() => pc1.close());
    t.add_cleanup(() => pc2.close());

    const stream = await getNoiseStream({audio: true});
    t.add_cleanup(() => stream.getTracks().forEach(track => track.stop()));
    pc1.addTrack(stream.getTracks()[0], stream);

    await pc1.setLocalDescription(await pc1.createOffer());
    await pc2.setRemoteDescription(pc1.pendingLocalDescription);

    await pc2.setLocalDescription(await pc2.createAnswer());
    await pc1.setRemoteDescription(pc2.localDescription);

    // Both ends want to add video at the same time. pc2 rolls back.

    const stream2 = await getNoiseStream({video: true});
    t.add_cleanup(() => stream2.getTracks().forEach(track => track.stop()));
    pc2.addTrack(stream2.getTracks()[0], stream2);
    await pc2.setLocalDescription(await pc2.createOffer());
    assert_equals(pc2.getTransceivers().length, 2);
    assert_not_equals(pc2.getTransceivers()[1].sender.transport, null)
    await pc2.setRemoteDescription({type: "rollback"});
    assert_equals(pc2.getTransceivers().length, 2);
    // Rollback didn't touch audio transceiver and transport is intact.
    assert_not_equals(pc2.getTransceivers()[0].sender.transport, null)
    // Video transport got killed.
    assert_equals(pc2.getTransceivers()[1].sender.transport, null)
    const stream1 = await getNoiseStream({video: true});
    t.add_cleanup(() => stream1.getTracks().forEach(track => track.stop()));
    pc1.addTrack(stream1.getTracks()[0], stream1);
    await pc1.setLocalDescription(await pc1.createOffer());
    await pc2.setRemoteDescription(pc1.pendingLocalDescription);
  },"Rollback local offer to negotiated stable state and then apply remote offer");

  promise_test(async t => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection();
    t.add_cleanup(() => pc1.close());
    t.add_cleanup(() => pc2.close());

    const stream = await getNoiseStream({audio: true});
    t.add_cleanup(() => stream.getTracks().forEach(track => track.stop()));
    pc1.addTrack(stream.getTracks()[0], stream);

    await pc1.setLocalDescription(await pc1.createOffer());
    await pc2.setRemoteDescription(pc1.pendingLocalDescription);

    await pc2.setLocalDescription(await pc2.createAnswer());
    await pc1.setRemoteDescription(pc2.localDescription);

    // pc1 adds video and pc2 adds audio. pc2 rolls back.
    assert_equals(pc2.getTransceivers()[0].direction, "recvonly");
    const stream2 = await getNoiseStream({audio: true});
    t.add_cleanup(() => stream2.getTracks().forEach(track => track.stop()));
    pc2.addTrack(stream2.getTracks()[0], stream2);
    assert_equals(pc2.getTransceivers()[0].direction, "sendrecv");
    await pc2.setLocalDescription(await pc2.createOffer());
    assert_equals(pc2.getTransceivers()[0].direction, "sendrecv");
    await pc2.setRemoteDescription({type: "rollback"});
    assert_equals(pc2.getTransceivers().length, 1);
    // setLocalDescription didn't change direction. So direction remains "sendrecv"
    assert_equals(pc2.getTransceivers()[0].direction, "sendrecv");
    // Rollback didn't touch audio transceiver and transport is intact. Still can receive audio.
    assert_not_equals(pc2.getTransceivers()[0].receiver.transport, null)
    const stream1 = await getNoiseStream({video: true});
    t.add_cleanup(() => stream1.getTracks().forEach(track => track.stop()));
    pc1.addTrack(stream1.getTracks()[0], stream1);
    await pc1.setLocalDescription(await pc1.createOffer());
    await pc2.setRemoteDescription(pc1.pendingLocalDescription);
  },"Rollback local offer with audio direction change to negotiated stable state and then add video receiver");

  promise_test(async t => {
    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection();
    t.add_cleanup(() => pc1.close());
    t.add_cleanup(() => pc2.close());

    pc1.addTransceiver('video', {direction: 'sendonly'});
    pc2.addTransceiver('video', {direction: 'sendonly'});
    await pc1.setLocalDescription(await pc1.createOffer());
    const pc1FirstMid = pc1.getTransceivers()[0].mid;
    await pc2.setLocalDescription(await pc2.createOffer());
    const pc2FirstMid = pc2.getTransceivers()[0].mid;
    // I don't think it is mandated that this has to be true, but any implementation I know of would
    // have predictable mids (e.g. 0, 1, 2...) so pc1 and pc2 should offer with the same mids.
    assert_equals(pc1FirstMid, pc2FirstMid);
    await pc1.setRemoteDescription(pc2.pendingLocalDescription);
    // We've implicitly rolled back and the SRD caused a second transceiver to be created.
    // As such, the first transceiver's mid will now be null, and the second transceiver's mid will
    // match the remote offer.
    assert_equals(pc1.getTransceivers().length, 2);
    assert_equals(pc1.getTransceivers()[0].mid, null);
    assert_equals(pc1.getTransceivers()[1].mid, pc2FirstMid);
    // If we now do an offer the first transceiver will get a different mid than in the first
    // pc1.createOffer()!
    pc1.setLocalDescription(await pc1.createAnswer())
    await pc1.setLocalDescription(await pc1.createOffer());
    assert_not_equals(pc1.getTransceivers()[0].mid, pc1FirstMid);
  },"Two transceivers with same mids");
</script>
